# src/main/resources/swagger.yaml
openapi: 3.0.0
info:
  title: Doodle assignment
  description: >-
    This API provides the a way to retrieve polls.
  version: 0.0.1
paths:
  /polls:
    get:
      summary: Returns the user created polls
      operationId: pollsSearch
      parameters:
        - $ref: '#/components/parameters/text'
        - $ref: '#/components/parameters/fromDate'
        - $ref: '#/components/parameters/untilDate'
      responses:
        '200':
          description: return
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PollResponse'
      tags:
        - Contracts

components:

  parameters:

    text:
      name: text
      in: query
      description: "full text search"
      required: false
      schema:
        type: string

    fromDate:
      name: fromdate
      in: query
      description: "Search from this date, if not provided the search will start from first record"
      required: false
      schema:
        type: string
        format: date

    untilDate:
      name: untildate
      in: query
      description: "Search until this date, if not provided then it will search until the last record"
      required: false
      schema:
        type: string
        format: date

  schemas:
    PollResponse:
      properties:
        id:
          type: string
        adminKey:
          type: string
        latestChange:
          type: integer
          format: int64
        initiated:
          type: integer
          format: int64
        participantsCount:
          type: integer
        inviteesCount:
          type: integer
        type:
          type: string
        hidden:
          type: boolean
        preferencesType:
          type: string
        state:
          type: string
        locale:
          type: string
        title:
          type: string
        initiator:
            $ref: '#/components/schemas/InitiatorResponse'
        options:
          type: array
          items:
            $ref: '#/components/schemas/OptionsResponse'
        optionsHash:
          type: string
        participants:
          type: array
          items:
            $ref: '#/components/schemas/ParticipantResponse'

    InitiatorResponse:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        notify:
          type: boolean

    OptionsResponse:
      type: object
      properties:
        text:
          type: string
        available:
          type: boolean
        start:
          type: number
        allday:
          type: boolean
        date:
          type: number
        end:
          type: string
        startDate:
          type: string
        endDate:
          type: string

    ParticipantResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        preferences:
          type: array
          items:
            type: number